Added icons to the project.
What has changed:

Added the "O" of the logo as an icon for the project.

This adds originality to the website, because everyone can now see this icon in the tab and everyone is now able to get quickly back to the website.

Added "Favicons for iOS - Web Clip" and "Favicons for Android Chrome".

Added favicons for MacOS Safari and the Windows Metro.