Convert menus to a hierarchy
Up until now the OpenDataDay site has been a single page site.
Now that we're adding one event page for every year we're going to need
more menu items and it is useful to think of them existing in a hiearachy.
This helps us to accomodate a larger number of menu items
that will grow over time

Databags:
  - represent menus as a hierarchical structure
  - use longer descriptions for languages
    language menu now has more horizontal space to play with
    so lets fill it a bit more

Templates:
  - update menu rendering code